only unlink if file exists
it was giving me an error that the file not exists on the second call

+++ src/Cache/FileCache.php
@@ -175,8 +175,10 @@ class FileCache implements CacheInterface
 
         $path = $this->buildRelativePath($uri, $query);
         $cache_file_path = $path . $this->results_filename;
-
-        unlink($cache_file_path);
+        
+        if (file_exists($cache_file_path)) {
+            unlink($cache_file_path);
+        }
     }
 
     /**
